Fix issue with rendering pre-rendered and instant pages.
Seems like we had a typo in the webRequest.onCompleted even handler.

This caused plenty of issue but most significantly affected prerendered
and instant pages in chrome.

You can reproduce this error in multiple ways:

1. Open up chrome and type in guardian.co.uk but don't hit enter.
2. Wait 5 seconds. (this will ensure page is prerendered in the
background.
3. Now hit enter.

Notice the clacks icon doesn't show up.

Prerendered pages have a tabId that ends up raising an error on
pageAction.show and hide. I've added safe versions of them.

This problem is significant because Google has started to add the gnu
header and nobody has noticed because of the bug.

To check this out just search for terry pratchett on google with this
patch committed and you'll see the clacks popup.