Resolved a runtime error when tabs variable is undefined
During debugging, the `tabs` variable occasionally was undefined. This
is not likely to occur in regular usage, and indeed it is difficult to
imagine a scenario in which one could invoke this function without a
tabs parameter being supplied, but it was observed.

Either way, the null check is harmless, and prevents a nasty error in
those rare cases.