Resolved defect whereby Clacks icon was always illuminated
In Chrome version 48, a change was made that affected the behaviour
and appearance of all extensions. This resulted in the Clacks Overhead
extension showing an illuminated icon on all pages, regardless of the
presence of an `X-Clacks-Overhead` header or meta tag. This commit
resolves this issue.

Prior to the release of Chrome 48, extensions could use the
`chrome.pageAction.show` function to display an icon in the Chrome
omnibox. The Clacks Overhead extension made use of this API to show an
illuminated Clacks icon when an X-Clacks-Overhead response header or
meta tag was detected in the active tab. Unless explicitly activated via
this API call, Chrome would either omit the icon, or render it in a
deactivated, greyed-out state.

Chrome 48 moved these icons to the Chrome toolbar, visible by default.
Extension icons are now 'activated' - not greyed-out - for every page
that the extension has the ability to access. It is no longer the case
that an extension could inspect a page and programmatically decide
whether or not to show its icon, nor to rely on Chrome to 'grey out' the
icon in case of a lack of a call to `chrome.pageAction.show`.

These changes seem to have been made to communicate better the existence
and activity of extensions in the browser, presumably to increase user
awareness of what is running at a particular time. This is a probably
a good change overall, but unfortunately one that has been communicated
poorly, and has not been reflected in the developer documentation.

The best way to restore the previous behaviour seems to be to include a
separate set of icons representing the 'greyed-out' state, and to alter
the extension icon programmatically, rather than relying solely on
`chrome.pageAction.show`.

Note that omitting the call to `chrome.pageAction.show` altogether
results in the popup failing to render, so it is still necessary. It is
just no longer sufficient.

For more information on the Chrome change that caused this defect,
refer to the release notes at https://support.google.com/chrome/forum/AAAAP1KN0B0t9AHfd90OAE

The following discussions and defect reports may also be helpful:
https://bugs.chromium.org/p/chromium/issues/detail?id=597657
https://bugs.chromium.org/p/chromium/issues/detail?id=970908