hotfix: update to Manifest v3
this is a hotfix to make the extension continue to work with the new
chrome store manifest change requirements.

it (unfortunately) does not include the popup window and other features,
but it does work on websites with the right header, and should allow
people to seamlessly update.

given that chrome is starting to break (see [this](https://www.reddit.com/r/discworld/comments/1g34hcz/chrome_just_updated_in_the_worst_possible_way/) reddit thread), this
is a quick fix so that the extension can be quickly repackaged.

fixes #14